requirements.txt
asgiref==3.8.1
attrs==25.1.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
defusedxml==0.8.0rc2
Django==3.2.25
django-filter==23.5
django-mail-templated==2.6.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
black
flake8
pytest
pytest-django
autopep8
djoser
Faker
isort